// src/calendars/enum.ts
export enum CALENDAR {
  GREGORIAN = "gregorian",
  BUDDHIST = "buddhist",
  HEBREW = "hebrew",
  INDIAN = "indian",
  COPTIC = "coptic",
  ETHIOPIC = "ethiopic",
  ETHIOPIC_AMETE_ALEM = "ethiopic-amete-alem",
  ISLAMIC_CIVIL = "islamic-civil",
  ISLAMIC_TABULAR = "islamic-tabular",
  ISLAMIC_UMALQURA = "islamic-umalqura",
  JAPANESE = "japanese",
  PERSIAN = "persian",
  TAIWAN = "roc", // Republic of China
}